pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "patch_analysis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.10.*"
dependencies = [
    "accelerate>=1.12.0",
    "bitsandbytes>=0.49.1",
    "datasets>=4.5.0",
    "diffusers>=0.36.0",
    "ipykernel>=7.1.0",
    "matplotlib>=3.10.8",
    "numpy>=2.2.6",
    "pandas>=2.3.3",
    "peft>=0.18.1",
    "pillow>=12.1.0",
    "protobuf>=6.33.4",
    "safetensors>=0.7.0",
    "scipy>=1.15.3",
    "sentencepiece>=0.2.1",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
    "transformers<5",
]
